Extract deserialize for share libs usage
